==> 0_Scrape_ParseHub.ipynb <==
# Using SQL + Pandas
import pandas as pd
import numpy as np

# Data Visualization
from plotnine import *
import seaborn as sns
import matplotlib.pyplot as plt

# Ignore warnings
import warnings
warnings.filterwarnings('ignore')

# Read in the data
DC_urls = pd.read_csv("Data/ParseHub/DC_urls.csv")

# Select every 10th row
DC_urls_sample = DC_urls.iloc[::10,:]
DC_urls_sample

# Export DC_urls_sample to a CSV file
DC_urls_sample.to_csv('Data/ParseHub/DC_urls_sample.csv',index=False)

# Read in the data
NYC_urls = pd.read_csv("Data/ParseHub/NYC_urls.csv")

# Select every 10th row
NYC_urls_sample = NYC_urls.iloc[::10,:]

# Export NYC_urls_sample to a CSV file
NYC_urls_sample.to_csv('Data/ParseHub/NYC_urls_sample.csv',index=False)

# Read in the data
CHI_urls = pd.read_csv("Data/ParseHub/CHI_urls.csv")

# Select every 5th row
CHI_urls_sample = CHI_urls.iloc[::5,:]

# Export CHI_urls_sample to a CSV file
CHI_urls_sample.to_csv('Data/ParseHub/CHI_urls_sample.csv',index=False)

# Read in the data
SF_urls = pd.read_csv("Data/ParseHub/SF_urls.csv")

# Select every 3rd row
SF_urls_sample = SF_urls.iloc[::3,:]

# Export CHI_urls_sample to a CSV file
SF_urls_sample.to_csv('Data/ParseHub/SF_urls_sample.csv',index=False)

# Read in the data
LA_urls = pd.read_csv("Data/ParseHub/LA_urls.csv")

# Select every 10th row
LA_urls_sample = LA_urls.iloc[::10,:]

# Export CHI_urls_sample to a CSV file
LA_urls_sample.to_csv('Data/ParseHub/LA_urls_sample.csv',index=False)
