==> 0_Scrape_Wiki.ipynb <==
## Preparation

# Data processing
import pandas as pd
import numpy as np
import re

# Scraping web content
import requests # For downloading the website
from bs4 import BeautifulSoup # For parsing the website
import time # To put the system to sleep
import random # For random numbers

# Ignore warnings
import warnings
warnings.filterwarnings('ignore')

## Scrape Wiki DC

# Wiki DC Michelin url
url = 'https://en.wikipedia.org/wiki/List_of_Michelin_starred_restaurants_in_Washington,_D.C.'

# Download the webpage
page = requests.get(url)
page.status_code # 200 == Connection

# Parse the content
soup = BeautifulSoup(page.content,'html.parser')

# Title of the wiki page
soup.title.string

# Get the right table
table = soup.find('table',{'class':'wikitable sortable'})

# Get the rows
rows = table.find_all('tr')
len(rows)

# Get the table header attributes
header = ([th.text.rstrip() for th in rows[0].find_all('th')] + 
          [re.sub(r'\(.*?\)\[.*?\]','',td.text).rstrip() for td in rows[0].find_all('td')])
header

# Build a function to extract star rating from an individual cell
def get_star(cell):
    """
    This is a function that takes cell as an input and output its value.
    """
    try:
        value = cell.find_all('img')[0]['alt']
    except IndexError:
        value = np.nan
    
    return value

# Create an empty DataFrame
Wiki_DC = pd.DataFrame(columns=header)

# Scrape and save to DataFrame
for row in rows[1:]:
    
    cells = row.find_all('td')
    
    Wiki_DC = Wiki_DC.append({header[0]:cells[0].text.rstrip(),
                              header[1]:cells[1].text.rstrip(),
                              header[2]:get_star(cells[2]),
                              header[3]:get_star(cells[3]),
                              header[4]:get_star(cells[4]),
                              header[5]:get_star(cells[5])},
                             ignore_index=True)

Wiki_DC

# Export to a CSV file
Wiki_DC.to_csv('Data/Wikipedia/Wiki_DC.csv',index=False)

## Scrape Wiki Chicago

# Wiki Chicago Michelin url
url = 'https://en.wikipedia.org/wiki/List_of_Michelin_starred_restaurants_in_Chicago'

# Download the webpage
page = requests.get(url)
page.status_code # 200 == Connection

# Parse the content
soup = BeautifulSoup(page.content,'html.parser')

# Title of the wiki page
soup.title.string

# Get the right table
table = soup.find('table',{'class':'wikitable sortable'})

# Get the rows
rows = table.find_all('tr')
len(rows)

# Get the table header attributes
header = [re.sub(r'\(.*?\)\[.*?\]','',th.text).rstrip() for th in rows[0].find_all('th')]
header

# Get the length of the header
len(header)

# Update get_star function to a deluxe edition
def get_star_deluxe(cells=None,i=None):
    """
    This is a a deluxe edition of the get_star function.
    """
    try:
        value = cells[i].find_all('img')[0]['alt']
    except IndexError:
        value = np.nan
    
    return value

# Create an empty DataFrame
Wiki_Chicago = pd.DataFrame(columns=header)

# Scrape and save to DataFrame
for row in rows[1:]:
    
    cells = row.find_all('td')
    
    Wiki_Chicago = Wiki_Chicago.append({header[0]:cells[0].text.rstrip(),
                                        header[1]:cells[1].text.rstrip(),
                                        header[2]:get_star_deluxe(cells,2),
                                        header[3]:get_star_deluxe(cells,3),
                                        header[4]:get_star_deluxe(cells,4),
                                        header[5]:get_star_deluxe(cells,5),
                                        header[6]:get_star_deluxe(cells,6),
                                        header[7]:get_star_deluxe(cells,7),
                                        header[8]:get_star_deluxe(cells,8),
                                        header[9]:get_star_deluxe(cells,9),
                                        header[10]:get_star_deluxe(cells,10),
                                        header[11]:get_star_deluxe(cells,11)},
                             ignore_index=True)

Wiki_Chicago

# Export to a CSV file
Wiki_Chicago.to_csv('Data/Wikipedia/Wiki_Chicago.csv',index=False)

## Scrape Wiki NY

# Wiki NY Michelin url
url = 'https://en.wikipedia.org/wiki/List_of_Michelin_starred_restaurants_in_New_York_City'

# Download the webpage
page = requests.get(url)
page.status_code # 200 == Connection

# Parse the content
soup = BeautifulSoup(page.content,'html.parser')

# Title of the wiki page
soup.title.string

# Get the right table
table = soup.find('table',{'class':'wikitable sortable'})

# Get the rows
rows = table.find_all('tr')
len(rows)

# Get the table header attributes
pat1 = r'\(.*?\)\[.*?\]\[.*?\]'
pat2 = r'\(.*?\)\[.*?\]'
com_pat = r'|'.join((pat1, pat2))
header = [re.sub(com_pat,'',th.text).rstrip() for th in rows[0].find_all('th')]
header

# Get the length of the header
len(header)

# Create an empty DataFrame
Wiki_NY = pd.DataFrame(columns=header)

# Scrape and save to DataFrame
for row in rows[1:]:
    
    cells = row.find_all('td')
    
    Wiki_NY = Wiki_NY.append({header[0]:cells[0].text.rstrip(),
                              header[1]:cells[1].text.rstrip(),
                              header[2]:get_star_deluxe(cells,2),
                              header[3]:get_star_deluxe(cells,3),
                              header[4]:get_star_deluxe(cells,4),
                              header[5]:get_star_deluxe(cells,5),
                              header[6]:get_star_deluxe(cells,6),
                              header[7]:get_star_deluxe(cells,7),
                              header[8]:get_star_deluxe(cells,8),
                              header[9]:get_star_deluxe(cells,9),
                              header[10]:get_star_deluxe(cells,10),
                              header[11]:get_star_deluxe(cells,11),
                              header[12]:get_star_deluxe(cells,12),
                              header[13]:get_star_deluxe(cells,13),
                              header[14]:get_star_deluxe(cells,14),
                              header[15]:get_star_deluxe(cells,15),
                              header[16]:get_star_deluxe(cells,16)},
                             ignore_index=True)

Wiki_NY

# Export to a CSV file
Wiki_NY.to_csv('Data/Wikipedia/Wiki_NY.csv',index=False)

## Scrape Wiki SF

# Wiki SF Michelin url
url = 'https://en.wikipedia.org/wiki/List_of_Michelin_starred_restaurants_in_San_Francisco_Bay_Area'

# Download the webpage
page = requests.get(url)
page.status_code # 200 == Connection

# Parse the content
soup = BeautifulSoup(page.content,'html.parser')

# Title of the wiki page
soup.title.string

# Get the right table
table = soup.find('table',{'class':'wikitable sortable'})

# Get the rows
rows = table.find_all('tr')
len(rows)

# Get the table header attributes
header = [re.sub(r'\(.*?\)\[.*?\]','',th.text).rstrip() for th in rows[0].find_all('th')]
header

# Get the length of the header
len(header)

# Create an empty DataFrame
Wiki_SF = pd.DataFrame(columns=header)

# Scrape and save to DataFrame
for row in rows[1:]:
    
    cells = row.find_all('td')
    
    Wiki_SF = Wiki_SF.append({header[0]:cells[0].text.rstrip(),
                              header[1]:cells[1].text.rstrip(),
                              header[2]:get_star(cells[2])},
                             ignore_index=True)

Wiki_SF

# Export to a CSV file
Wiki_SF.to_csv('Data/Wikipedia/Wiki_SF.csv',index=False)

## Scrape Wiki LA

# Wiki LA Michelin url
url = 'https://en.wikipedia.org/wiki/List_of_Michelin_starred_restaurants_in_Los_Angeles'

# Download the webpage
page = requests.get(url)
page.status_code # 200 == Connection

# Parse the content
soup = BeautifulSoup(page.content,'html.parser')

# Title of the wiki page
soup.title.string

# Get the right table
table = soup.find('table',{'class':'wikitable sortable'})

# Get the rows
rows = table.find_all('tr')
len(rows)

# Get the table header attributes
header = [re.sub(r'\[.*?\]','',th.text).rstrip() for th in rows[0].find_all('th')]
header

# Get the length of the header
len(header)

# Create an empty DataFrame
Wiki_LA = pd.DataFrame(columns=header)

# Scrape and save to DataFrame
for row in rows[1:]:
    
    cells = row.find_all('td')
    
    Wiki_LA = Wiki_LA.append({header[0]:cells[0].text.rstrip(),
                              header[1]:cells[1].text.rstrip(),
                              header[2]:get_star_deluxe(cells,2),
                              header[3]:get_star_deluxe(cells,3),
                              header[4]:get_star_deluxe(cells,4),
                              header[5]:get_star_deluxe(cells,5)},
                             ignore_index=True)

Wiki_LA

# Export to a CSV file
Wiki_LA.to_csv('Data/Wikipedia/Wiki_LA.csv',index=False)
