==> 0_X_Scrape_Yelp.ipynb <==
## Preparation

# Data processing
import pandas as pd

# Scraping web content
import requests # For downloading the website
from bs4 import BeautifulSoup # For parsing the website
import time # To put the system to sleep
import random # For random numbers

# Ignore warnings
import warnings
warnings.filterwarnings('ignore')

## Extract Relevant Links

# Yelp DC restaurants url
DC_url = 'https://www.yelp.com/search?find_desc=Restaurants&find_loc=Washington%2C%20DC'

# Download the webpage
DC_page = requests.get(DC_url)
DC_page.status_code # 200 == Connection

# Parse the content
DC_soup = BeautifulSoup(DC_page.content,'html.parser')

# Extract page urls
DC_urls = set([DC_url])
DC_page_no = 24

for i in range(10, DC_page_no*10, 10):
    DC_urls.update([DC_url+'&start='+str(i)])

DC_urls

# Extract relevant links for one page
links = set()

for tag in DC_soup.select('span > a'):
    href = tag.attrs.get('href')
    if '?osq=Restaurants' in href and 'https:' not in href:
        links.update(['https://www.yelp.com'+href])

links

# Build a scraper to extract relevant links for multiple pages
def page_scraper(urls=None,sleep=3):
    """
    Scrape Yelp urls to extract relevant links.

    Args:
        urls (list): list of Yelp urls.
        sleep (int): integer value specifying how long the machine should be put to sleep (random uniform); defaults to 3.

    Returns:
        set: set containing all relevant restaurant links.
    """
    links = set()

    for url in urls:

        # Keep track of where we are at
        print(url)

        # Download the webpage
        page = requests.get(url)

        # If a connection was reached
        if page.status_code == 200:

                # Parse
                soup = BeautifulSoup(page.content,'html.parser')

                for tag in soup.select('span > a'):
                    href = tag.attrs.get('href')
                    if '?osq=Restaurants' in href and 'https:' not in href:
                        links.update(['https://www.yelp.com'+href])

        # Put the system to sleep for a random draw of time
        time.sleep(random.uniform(0,sleep))

    # Return data
    return links

# Scrape Yelp DC page
DC_links = page_scraper(urls=DC_urls)

# View DC_links
DC_links

# Check number of DC_links
len(DC_links)

## Scrape Yelp DC

# Create an empty DataFrame
Yelp_DC = pd.DataFrame(columns=['business_name','rating',
                                'review_count','price_category'])

for tag in DC_soup.select('[class*=container]'):
    if tag.find('h4'):
        Yelp_DC = Yelp_DC.append({'business_name':tag.select('h4')[0].get_text(),
                                  'rating':tag.select('[class*=stars]')[0]['aria-label'],
                                  'review_count':tag.select('[class*=reviewCount]')[0].get_text(),
                                  'price_category':tag.select('[class*=priceCategory]')[0].get_text()},
                                 ignore_index=True)

Yelp_DC

# Build a scraper for Yelp links
def Yelp_scraper(url=None):
    """
    Scrape a Yelp url to extract business information.

    Args:
        url (str): string of a Yelp url.

    Returns:
        DataFrame: frame containing business information of the url.
    """
    # Download the webpage
    page = requests.get(url)

    # If a connection was reached
    if page.status_code == 200:

        # Parse
        soup = BeautifulSoup(page.content,'html.parser')

        # Create an empty DataFrame
        info = pd.DataFrame(columns=['business_name','rating','review_count','price_category'])

        for tag in soup.select('[class*=container]'):
            if tag.find('h4'):
                info = info.append({'business_name':tag.select('h4')[0].get_text(),
                                    'rating':tag.select('[class*=stars]')[0]['aria-label'],
                                    'review_count':tag.select('[class*=reviewCount]')[0].get_text(),
                                    'price_category':tag.select('[class*=priceCategory]')[0].get_text()},
                                   ignore_index=True)

    # Return data
    return info

# Scrape DC_url
Yelp_scraper(url=DC_url)

# Build a function to scrape multiple Yelp links
def link_scrape(urls=None,sleep=3):
    """
    Scrape multiple Yelp links.

    Args:
        urls (list): list of Yelp urls.
        sleep (int): integer value specifying how long the machine should be put to sleep (random uniform); defaults to 3.

    Returns:
        DataFrame: frame containing business information of all urls.
    """
    dat = pd.DataFrame([])

    for url in urls:

        print(url) # Keep track of where we are at

        try:

            # Scrape the content
            dat = dat.append(Yelp_scraper(url))

            # Put the system to sleep for a random draw of time
            time.sleep(random.uniform(0,sleep))

        except ImportError:
            pass

    dat = dat.reset_index(drop=True)
    return dat

# Scrape DC_urls
DC_Yelp = link_scrape(urls=DC_urls)

# View a random sample of DC_yelp
DC_Yelp.sample(10)

# Export DC_Yelp to a CSV file
DC_Yelp.to_csv('Data/Yelp/DC_Yelp.csv', index = False)

## Scrape Yelp Chicago

# Yelp Chicago restaurants url
Chicago_url = 'https://www.yelp.com/search?find_desc=Restaurants&find_loc=Chicago%2C%20IL'

# Download the webpage
Chicago_page = requests.get(Chicago_url)
Chicago_page.status_code # 200 == Connection

# Parse the content
Chicago_soup = BeautifulSoup(Chicago_page.content,'html.parser')

# Extract page urls
Chicago_urls = set([Chicago_url])
Chicago_page_no = 24

for i in range(10, Chicago_page_no*10, 10):
    Chicago_urls.update([Chicago_url+'&start='+str(i)])

Chicago_urls

# Scrape Chicago_urls
Chicago_Yelp = link_scrape(urls=Chicago_urls)

# View a random sample of Chicago_yelp
Chicago_Yelp.sample(10)

# Export Chicago_Yelp to a CSV file
Chicago_Yelp.to_csv('Data/Yelp/Chicago_Yelp.csv', index = False)

## Scrape Yelp New York

# Yelp NY restaurants url
NY_url = 'https://www.yelp.com/search?find_desc=Restaurants&find_loc=New%20York%2C%20NY'

# Download the webpage
NY_page = requests.get(NY_url)
NY_page.status_code # 200 == Connection

# Parse the content
NY_soup = BeautifulSoup(NY_page.content,'html.parser')

# Extract page urls
NY_urls = set([NY_url])
NY_page_no = 24

for i in range(10, NY_page_no*10, 10):
    NY_urls.update([NY_url+'&start='+str(i)])

NY_urls

# Scrape NY_urls
NY_Yelp = link_scrape(urls=NY_urls)

# View a random sample of NY_yelp
NY_Yelp.sample(10)

# Export NY_Yelp to a CSV file
NY_Yelp.to_csv('Data/Yelp/NY_Yelp.csv', index = False)

## Scrape Yelp San Francisco Bay Area

# Yelp SF restaurants url
SF_url = 'https://www.yelp.com/search?find_desc=Restaurants&find_loc=San%20Francisco%20Bay%20Area%2C%20CA'

# Download the webpage
SF_page = requests.get(SF_url)
SF_page.status_code # 200 == Connection

# Parse the content
SF_soup = BeautifulSoup(NY_page.content,'html.parser')

# Extract page urls
SF_urls = set([SF_url])
SF_page_no = 24

for i in range(10, SF_page_no*10, 10):
    SF_urls.update([SF_url+'&start='+str(i)])

SF_urls

# Scrape SF_urls
SF_Yelp = link_scrape(urls=SF_urls)

# View a random sample of SF_yelp
SF_Yelp.sample(10)

# Export SF_Yelp to a CSV file
SF_Yelp.to_csv('Data/Yelp/SF_Yelp.csv', index = False)

## Scrape Yelp Los Angeles

# Yelp LA restaurants url
LA_url = 'https://www.yelp.com/search?find_desc=Restaurants&find_loc=Los%20Angeles%2C%20CA'

# Download the webpage
LA_page = requests.get(LA_url)
LA_page.status_code # 200 == Connection

# Parse the content
LA_soup = BeautifulSoup(LA_page.content,'html.parser')

# Extract page urls
LA_urls = set([LA_url])
LA_page_no = 24

for i in range(10, LA_page_no*10, 10):
    LA_urls.update([LA_url+'&start='+str(i)])

LA_urls

# Scrape LA_urls
LA_Yelp = link_scrape(urls=LA_urls)

# View a random sample of LA_yelp
LA_Yelp.sample(10)

# Export LA_Yelp to a CSV file
LA_Yelp.to_csv('Data/Yelp/LA_Yelp.csv', index = False)
