==> 2_Final_Merge.ipynb <==
# Using SQL + Pandas
import pandas as pd
import numpy as np

# Data Visualization
from plotnine import *
import seaborn as sns
import matplotlib.pyplot as plt

# Ignore warnings
import warnings
warnings.filterwarnings('ignore')

# Read in the data
DC = pd.read_csv("Data/DC.csv")
NY = pd.read_csv("Data/NY.csv")
CH = pd.read_csv("Data/CH.csv")
CA = pd.read_csv("Data/CA.csv")

# Row bind all data
dat = pd.concat([DC,NY,CH,CA],sort=False)

# View the head of the data
dat.head()

# View the shape of the data
dat.shape

# View the data type of each column
dat.dtypes

# Count NaNs in each column
dat.isna().sum()

# Drop NaNs
dat.dropna(subset=['price_range','category'],inplace=True)
dat.shape

import ast
dat.category = dat.category.apply(ast.literal_eval)

cat_dummies = (pd
               .get_dummies(dat.category.apply(pd.Series),
                            prefix='',
                            prefix_sep='')
               .sum(level=0, axis=1))

cat_dummies.head()

dat = pd.concat([dat,cat_dummies],axis=1,sort=False)

cat_dummies.columns[300:]

drop = ['Airport Lounges', 'Antiques', 'Apartments',
       'Arcades','Art Galleries', 'Art Supplies',
       'Banks & Credit Unions', 'Bespoke Clothing', 'Bikes',
       'Blow Dry/Out Services', 'Boating', 'Bookstores', 'Botanical Gardens',
       'Bowling', 'Bridal', 'Casinos', 'Cinema',
       'Comedy Clubs', 'Convenience Stores', 'Cooking Classes', 'Cooking Schools',
       'Cosmetics & Beauty Supply', 'Dance Clubs', 'Day Spas',
       'Do-It-Yourself Food', 'Drugstores', 'Electronics Repair',
       'Escape Games', 'Festivals', 'Florists',
       'Framing', 'Fur Clothing',
       'Furniture Stores', 'Gas Stations', 'Gold Buyers', 'Golf',
       'Grocery', 'Hair Loss Centers', 'Hair Removal',
       'Hair Salons', 'Hair Stylists', 'Health Markets',
       'Hotels', 'International Grocery', 'Jewelry',
       'Karaoke', 'Landmarks & Historical Buildings', 'Laotian', 'Latin American',
       'Lingerie', 'Magicians', 
       "Men's Clothing", 'Museums', 'Music & DVDs', 'Music Venues',
       'Musical Instruments & Teachers', 'Nail Salons', 'Organic Stores',
       'Performing Arts', 'Pet Stores', 'Plus Size Fashion',
       'Pool Halls', 'Pubs', 'Resorts', 'Rugs',
       'Seafood Markets', 'Sewing & Alterations', 'Shopping Centers', 
       'Speakeasies', 'Tattoo', 'Tobacco Shops', 'Tours', 'Used', 'Vape Shops',
       'Venues & Event Spaces', 'Vinyl Records',
       "Women's Clothing", 'Accessories', 'Art Museums',
       'Bed & Breakfast', 'Beverage Store',
       'Cabaret', 'Cultural Center', 'Electronics', 'Event Planning & Services',
       'Eyelash Service', 'Floral Designers', 'Flowers & Gifts', 
       'Food Tours', 'Formal Wear', 'Furniture Rental', 'Gift Shops', 
       'Head Shops', 'Home Decor', 'Ice Delivery',
       'Internet Cafes', 'Libraries', 'Makeup Artists', 'Mattresses',
       'Meaderies', 'Mobile Phone Repair', 'Party & Event Planning', 
       'Personal Chefs', 'Personal Shopping', 'Pop-Up Restaurants', 
       'Public Markets', 'Recording & Rehearsal Studios',
       'Shopping', 'Tabletop Games', 'Team Building Activities',
       'Trivia Hosts', 'Vacation Rentals', 
       'Veterans Organizations', 'Vintage & Consignment', 'Wedding Planning',
       'Wholesale Stores', 'Yoga',
       'Aircraft Repairs', 'Art Tours', 'Beer Tours',
       'Bike Repair/Maintenance', 'Boat Tours', 'Books', 
       'Herbs & Spices', 'Kids Activities',
       'Mini Golf', 'Musicians', 'Office Equipment', 'Piercing',
       'Skin Care', 'Swimming Pools',
       'Used Bookstore', 'Walking Tours', 'Mags', 'Music & Video',
       'Breweries', 'Brewpubs', 'Candy Stores', 'Cheese Shops', 
       'Custom Cakes', 'Distilleries', 'Dive Bars', 
       'Food Court', 'Food Delivery Services', 'Food Stands',
       'Food Trucks', 'Fruits & Veggies', 'Gastropubs', 'Gay Bars',
       'Hookah Bars', 'Imported Food', 'Irish Pub', 'Jazz & Blues', 'Meat Shops',
       'Patisserie/Cake Shop', 'Piano Bars', 'Sports Bars', 
       'Street Vendors', 'Supper Clubs', 'Australian', 'Bartenders', 
       'Dinner Theater', 'Social Clubs', 'Sports Clubs']

for i in drop:
    dat = dat[dat[i] != 1]

dat.drop(columns=drop,inplace=True)

dat.shape

# Export to a CSV file
dat.to_csv('Data/restaurant_data.csv',index=False)
