==> 2_X_Merge_All.ipynb <==
## Preparation

# Using SQL + Pandas
import pandas as pd
import numpy as np

# Data Visualization
from plotnine import *
import seaborn as sns
import matplotlib.pyplot as plt

# Ignore warnings
import warnings
warnings.filterwarnings('ignore')

# Read in the data
DC_Yelp = pd.read_csv("Data/Yelp/DC_Yelp_cleaned.csv")
Chicago_Yelp = pd.read_csv("Data/Yelp/Chicago_Yelp_cleaned.csv")
NY_Yelp = pd.read_csv("Data/Yelp/NY_Yelp_cleaned.csv")
SF_Yelp = pd.read_csv("Data/Yelp/SF_Yelp_cleaned.csv")
LA_Yelp = pd.read_csv("Data/Yelp/LA_Yelp_cleaned.csv")

DC_Wiki = pd.read_csv("Data/Wikipedia/Wiki_DC.csv")
Chicago_Wiki = pd.read_csv("Data/Wikipedia/Wiki_Chicago.csv")
NY_Wiki = pd.read_csv("Data/Wikipedia/Wiki_NY.csv")
SF_Wiki = pd.read_csv("Data/Wikipedia/Wiki_SF.csv")
LA_Wiki = pd.read_csv("Data/Wikipedia/Wiki_LA.csv")

## Merge DC data

DC_Yelp

DC_Wiki

list(DC_Wiki.columns)[:-1]

DC_YW = (pd
         .merge(left = DC_Yelp,
                right = DC_Wiki,
                how = "left",
                left_on = "business_name",
                right_on="Name")
         .drop(list(DC_Wiki.columns)[:-1],axis=1)
         .rename(columns={"2020":"michelin_star"}))

DC_YW

DC_YW['michelin_star'] = DC_YW.michelin_star.fillna(0)
DC_YW['michelin_star'] = DC_YW.michelin_star.replace('1 Michelin star',1)
DC_YW['michelin_star'] = DC_YW.michelin_star.replace('2 Michelin stars',2)
DC_YW['michelin_star'] = DC_YW.michelin_star.replace('3 Michelin stars',3)
DC_YW['region'] = 'Washington DC'

DC_YW

DC_YW.query('michelin_star != 0')

## Merge Chicago data

Chicago_YW = (pd
              .merge(left = Chicago_Yelp,
                     right = Chicago_Wiki,
                     how = "left",
                     left_on = "business_name",
                     right_on="Name")
              .drop(list(Chicago_Wiki.columns)[:-1],axis=1)
              .rename(columns={"2020":"michelin_star"}))

Chicago_YW['michelin_star'] = Chicago_YW.michelin_star.fillna(0)
Chicago_YW['michelin_star'] = Chicago_YW.michelin_star.replace('1 Michelin star',1)
Chicago_YW['michelin_star'] = Chicago_YW.michelin_star.replace('2 Michelin stars',2)
Chicago_YW['michelin_star'] = Chicago_YW.michelin_star.replace('3 Michelin stars',3)
Chicago_YW['region'] = 'Chicago'

Chicago_YW

Chicago_YW.query('michelin_star != 0')

## Merge NY data

NY_YW = (pd
         .merge(left = NY_Yelp,
                right = NY_Wiki,
                how = "left",
                left_on = "business_name",
                right_on="Name")
         .drop(list(NY_Wiki.columns)[:-1],axis=1)
         .rename(columns={"2020":"michelin_star"}))

NY_YW['michelin_star'] = NY_YW.michelin_star.fillna(0)
NY_YW['michelin_star'] = NY_YW.michelin_star.replace('1 Michelin star',1)
NY_YW['michelin_star'] = NY_YW.michelin_star.replace('2 Michelin stars',2)
NY_YW['michelin_star'] = NY_YW.michelin_star.replace('3 Michelin stars',3)
NY_YW['region'] = 'New York City'

NY_YW

NY_YW.query('michelin_star != 0')

## Merge SF data

SF_YW = (pd
         .merge(left = SF_Yelp,
                right = SF_Wiki,
                how = "left",
                left_on = "business_name",
                right_on="Name")
         .drop(list(SF_Wiki.columns)[:-1],axis=1)
         .rename(columns={"2019":"michelin_star"}))

SF_YW['michelin_star'] = SF_YW.michelin_star.fillna(0)
SF_YW['michelin_star'] = SF_YW.michelin_star.replace('1 Michelin star',1)
SF_YW['michelin_star'] = SF_YW.michelin_star.replace('2 Michelin stars',2)
SF_YW['michelin_star'] = SF_YW.michelin_star.replace('3 Michelin stars',3)
SF_YW['region'] = 'San Francisco'

SF_YW

SF_YW.query('michelin_star != 0')

## Merge LA Data

LA_YW = (pd
         .merge(left = LA_Yelp,
                right = LA_Wiki,
                how = "left",
                left_on = "business_name",
                right_on="Name")
         .drop(list(LA_Wiki.columns)[:-1],axis=1)
         .rename(columns={"2019":"michelin_star"}))

LA_YW['michelin_star'] = LA_YW.michelin_star.fillna(0)
LA_YW['michelin_star'] = LA_YW.michelin_star.replace('1 Michelin star',1)
LA_YW['michelin_star'] = LA_YW.michelin_star.replace('2 Michelin stars',2)
LA_YW['michelin_star'] = LA_YW.michelin_star.replace('3 Michelin stars',3)
LA_YW['region'] = 'Los Angeles'

LA_YW

LA_YW.query('michelin_star != 0')

## Stack All Data

Yelp_Wiki = pd.concat([DC_YW,Chicago_YW,NY_YW,SF_YW,LA_YW],ignore_index=True)
Yelp_Wiki

Yelp_Wiki.dtypes

# Read in the Michelin data
Michelin = pd.read_csv("Data/Michelin/Michelin.csv")
Michelin = Michelin.drop(['region','url'],axis=1)
Michelin

df = (pd
      .merge(left = Yelp_Wiki,
             right = Michelin,
             how = "left",
             left_on = "business_name",
             right_on="name")
      .drop('name',axis=1))

df['michelin_guide'] = df.michelin_guide.fillna(0).astype(int)

df

# Export to a CSV file
df.to_csv('Data/Merged_Data.csv',index=False)
